fix: correccion integracion backend-frontend

+++ Code/Server/src/Application/Features/Gavetero/GaveteroDto.cs
@@ -6,7 +6,7 @@ public class GaveteroDto
     public string? Nombre { get; set; }
     public string? Tipo { get; set; }
     public string? NombreMueble { get; set; }
-    public decimal? Longitud { get; set; }
-    public decimal? Profundidad { get; set; }
-    public decimal? Altura { get; set; }
+    public double? Longitud { get; set; }
+    public double? Profundidad { get; set; }
+    public double? Altura { get; set; }
 }
